move scripts to a dedicated folder
